ndisc: Filter received unicast targets
If the solicit to a specific target gets dropped or doesn't receive a
response fast enough, advertisements from other routers can be received
first. When we set an explicit unicast target, it is generally
unexpected to see these advertisements output to console. Filter out
received advertisements if the target is a unicast IPv6 address.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>