Revert "ndisc: Filter received unicast targets"
This reverts commit 835ff19962b6ef59f70ac99bd952b7375ad54732.

There are no requirements for responders to match the destination
address of the solitication in their advertisements. And indeed, a
number of implementations are known *not* to do that matching.

This is admittedly annoying, especially for router advertisements,
we cannot filter advertisements based on much any criterion, but we
cannot be more restrictive than the specification, especially when it
is known to break actual interoperability.

(For neighbor advertisements, we can at least filter based on the
advertised address).